Start restructuring the io/trace.h to prepare for streaming trace
