+++ src/workerd/io/trace-common.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include <workerd/io/outcome.capnp.h>
+#include <workerd/io/worker-interface.capnp.h>
+
+namespace workerd {
+
+using LogLevel = rpc::Trace::Log::Level;
+using ExecutionModel = rpc::Trace::ExecutionModel;
+
+enum class PipelineLogLevel {
+  // WARNING: This must be kept in sync with PipelineDef::LogLevel (which is not in the OSS
+  //   release).
+  NONE,
+  FULL
+};
+
+}  // namespace workerd
